requirements.txt
slackclient>=2.9.1
slackeventsapi>=2.2.1
Flask>=1.1.2
idna~=2.10
certifi~=2020.6.20
chardet~=3.0.4
requests~=2.24.0
urllib3~=1.25.10
unleashed-py>=1.0.8